+++ nmanga/cmd.py
@@ -33,7 +33,7 @@ import click
 from ._metadata import __author__, __name__, __version__
 from .cli.archive import pack_releases, pack_releases_comment_archive, pack_releases_epub_mode
 from .cli.auto_split import auto_split
-from .cli.autolevel import autolevel, autolevel2, force_gray
+from .cli.autolevel import auto_posterize, autolevel, autolevel2, force_gray, posterize_simple
 from .cli.base import NMangaCommandHandler
 from .cli.config import cli_config
 from .cli.denoiser import denoiser, denoiser_trt, identify_denoise_candidates
@@ -116,6 +116,8 @@ main.add_command(timewizard_modify)
 main.add_command(autolevel)
 main.add_command(autolevel2)
 main.add_command(force_gray)
+main.add_command(posterize_simple)
+main.add_command(auto_posterize)
 main.add_command(denoiser)
 main.add_command(denoiser_trt)
 main.add_command(identify_denoise_candidates)